Update CHANGESET.md
Summary of changes from commit log; mostly focussed on "hard" changes like the package updates (condensed to latest version numbers); questionable on the `LICENSE` file rename (probably of no interest to the reader, but I followed principle of gae-init file renames in previous `CHANGESET.md` entries; you can easily drop this bullet); what might be missing is a highlight of the new features. Note that I changed the release date of 0.9.0 to "tbd" (= to be determined).